+++ docs/support/requirements_to_rst.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # requirements_to_rst.py
-# Copyright (c) 2013-2017 Pablo Acosta-Serafini
+# Copyright (c) 2013-2018 Pablo Acosta-Serafini
 # See LICENSE for details
 # pylint: disable=C0103,C0111,R0912,R0914,R0915
 
@@ -84,8 +84,7 @@ def op_to_words(item):
                 return suffix+item[2:]
             elif prefix in ['>', '<']:
                 return suffix+item[1:]
-            else:
-                return item[2:]+suffix
+            return item[2:]+suffix
     raise RuntimeError('Inequality not supported')
 
 
@@ -118,8 +117,7 @@ def ops_to_words(item):
             +' '+
             op_to_words(tokens[-1])
         )
-    else:
-        return ' and '.join([op_to_words(token) for token in tokens])
+    return ' and '.join([op_to_words(token) for token in tokens])
 
 
 def proc_requirements(mobj):