+++ sbin/gen_req_files.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # gen_req_files.py
-# Copyright (c) 2013-2017 Pablo Acosta-Serafini
+# Copyright (c) 2013-2018 Pablo Acosta-Serafini
 # See LICENSE for details
 # pylint: disable=C0111