+++ tests/test.py
@@ -1,5 +1,5 @@
 # test.py
-# Copyright (c) 2013-2017 Pablo Acosta-Serafini
+# Copyright (c) 2013-2018 Pablo Acosta-Serafini
 # See LICENSE for details
 # pylint: disable=C0111,R0201,R0903,W0212