+++ tests/test_pmisc.py
@@ -1,5 +1,5 @@
 # test_pmisc.py
-# Copyright (c) 2013-2017 Pablo Acosta-Serafini
+# Copyright (c) 2013-2018 Pablo Acosta-Serafini
 # See LICENSE for details
 # pylint: disable=C0111,W0401,W0614